// src/app/models/recruitment/applyaddress.ts
export class ReqaddressModel {
    constructor() {

    }

    company_code: string = "";
    applywork_code: string = "";
    applyaddress_id: string = "1";
    applyaddress_type: string = "";
    applyaddress_no: string = "";
    applyaddress_moo: string = "";
    applyaddress_soi: string = "";
    applyaddress_road: string = "";
    applyaddress_tambon: string = "";
    applyaddress_amphur: string = "";
    applyaddress_zipcode: string = "";

    applyaddress_tel: string = "";
    applyaddress_email: string = "";
    applyaddress_line: string = "";
    applyaddress_facebook: string = "";
    applyprovince_code: string = "";

    modified_by: string = "";
    modified_date: string = "";

    index: number = 0;
    select: boolean = false;
    applyaddress_ref: string = "";
}